requirements.txt
# Core dependencies for Digital Twin optimized Python OMB
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pyyaml>=6.0
aiohttp>=3.9.0
click>=8.1.0
rich>=13.0.0

# Data processing and visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# System monitoring
psutil>=5.9.0

# Template rendering
jinja2>=3.1.0

# Kafka client (Digital Twin optimized)
confluent-kafka>=2.3.0

# High-precision latency measurement (like original OMB)
hdrhistogram>=0.10.3